Migrate logic to auto register user into CreateTableHandler
